move nim version in issue template to the top (#24733)
The point is to move it out of the current place between "Description"
and "Current Output" as often these are related to each other and
including the nim version in the middle breaks the flow of reading. I
also thought of moving it below "Expected Output" but this felt too low
for a required field and also has the same problem of overshadowing the
remaining sections. Not sure if it being at the top is annoying in some
other way though.